Make ChannelId a truly sealed newtype.
